Add API Key authorization to access endpoint
Updated app.py to parse and validate API keys in
request headers. The lack of documentation for
FastAPI's Security() and built-in APIKey types
means that the API key will be read the old-
fashioned way.

This will be fixed once better documentation
and/or samples are available, but since Vercel
uses HTTPS and the key is held in an environment
variable, this is still fairly secure.